+++ src/app/(views)/viewer/components/SplatViewer.tsx
@@ -18,8 +18,14 @@ import { InfoPanel } from "./InfoPanel";
 import SceneItem from "../../../lib/definitions/SceneItem";
 
 interface SplatViewerProps {
+  sceneItem: {
+    id: number;
+    name: string;
+    description: string;
+    splatUrl: string;
+    videoUrl: string;
+  } | null;
   onClose: () => void;
-  sceneItem: SceneItem | null;
 }
 
 export default function SplatViewer({ onClose, sceneItem }: SplatViewerProps) {